# annotation/documentation/openapi.yaml
openapi: 3.0.2
info:
  title: Badgerdoc Annotation
  version: 0.1.5
paths:
  /annotation/{task_id}:
    post:
      tags:
      - Annotation
      - Annotation
      summary: Save annotation by user.
      description: |-
        Saves annotated doc (annotation) by user. For first save base
        revision of given annotated doc should be null. "Pages", "validated"
        and "failed_validation_pages" arrays should not be empty at the same
        time, because there will be nothing to save. Tenant in header is the
        name of the bucket in minIO, where annotated pages will be saved.
        Path to pages from bucket (tenant) in minIO is as follows -
        annotation / {job_id} / {file_id}. Job_id and file_id are received from
        task entity by provided task_id in path. Also, by this path in minIO there
        will be manifest.json file, that will contain "pages" field with hashes
        for pages of only latest revisions, "validated"/"failed_validation_pages"
        fields with numbers of validated/failed_validation pages respectively.
      operationId: post_annotation_by_user_annotation__task_id__post
      parameters:
      - required: true
        schema:
          title: Task Id
          type: integer
        example: 5
        name: task_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocForSaveSchema'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotatedDocSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
  /annotation/{job_id}/{file_id}:
    get:
      tags:
      - Annotation
      - Revision
      - Annotation
      summary: Get all users revisions (or pipeline revision) for particular pages.
      operationId: get_all_revisions_annotation__job_id___file_id__get
      parameters:
      - required: true
        schema:
          title: Job Id
          type: integer
        name: job_id
        in: path
      - required: true
        schema:
          title: File Id
          type: integer
        name: file_id
        in: path
      - required: true
        schema:
          title: Page Numbers
          minimum: 1.0
          minItems: 1
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
        name: page_numbers
        in: query
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get All Revisions Annotation  Job Id   File Id  Get
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/PageOutSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
    post:
      tags:
      - Annotation
      - Annotation
      summary: Save annotation by pipeline.
      description: |-
        Saves annotated doc (annotation) by pipeline. Base revision of
        given annotated doc should always be null. "Pages", "validated"
        and "failed_validation_pages" arrays should not be empty at the
        same time, because there will be nothing to save. Tenant in header
        is the name of the bucket in minIO, where annotated pages will
        be saved. Path to pages from bucket (tenant) in minIO is as follows -
        annotation / {job_id} / {file_id}. Also by this path in minIO there
        will be manifest.json file, that will contain "pages" field with hashes
        for pages of only latest revisions, "validated"/"failed_validation_pages"
        fields with numbers of validated/failed_validation pages respectively.
      operationId: post_annotation_by_pipeline_annotation__job_id___file_id__post
      parameters:
      - required: true
        schema:
          title: Job Id
          type: integer
        example: 3
        name: job_id
        in: path
      - required: true
        schema:
          title: File Id
          type: integer
        example: 4
        name: file_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocForSaveSchema'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnnotatedDocSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '400':
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BadRequestErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
  /annotation/{file_id}:
    get:
      tags:
      - Annotation
      - Jobs
      summary: Get all job_ids that have annotations for the given file_id.
      operationId: get_jobs_by_file_id_annotation__file_id__get
      parameters:
      - required: true
        schema:
          title: File Id
          type: integer
        example: 4
        name: file_id
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Jobs By File Id Annotation  File Id  Get
                type: array
                items:
                  $ref: '#/components/schemas/JobOutSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
  /annotation/{job_id}/{file_id}/latest_by_user:
    get:
      tags:
      - Annotation
      - Revision
      - Annotation
      summary: Get latest revision made by particular user (or by pipeline) for particular
        pages.
      operationId: get_latest_revision_by_user_annotation__job_id___file_id__latest_by_user_get
      parameters:
      - required: true
        schema:
          title: Job Id
          type: integer
        example: 3
        name: job_id
        in: path
      - required: true
        schema:
          title: File Id
          type: integer
        example: 4
        name: file_id
        in: path
      - required: true
        schema:
          title: Page Numbers
          minimum: 1.0
          minItems: 1
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
        example:
        - 1
        - 3
        - 4
        name: page_numbers
        in: query
      - required: false
        schema:
          title: User Id
          type: string
          format: uuid
        example: 1843c251-564b-4c2f-8d42-c61fdac369a1
        name: user_id
        in: query
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                title: Response Get Latest Revision By User Annotation  Job Id   File
                  Id  Latest By User Get
                type: object
                additionalProperties:
                  type: array
                  items:
                    $ref: '#/components/schemas/PageOutSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
  /annotation/{job_id}/{file_id}/{revision}:
    get:
      tags:
      - Annotation
      - Revision
      - Annotation
      summary: Get annotation for given revision.Info will be accumulated from first
        revision up togiven.
      operationId: get_annotations_up_to_given_revision_annotation__job_id___file_id___revision__get
      parameters:
      - required: true
        schema:
          title: Job Id
          type: integer
        example: 1
        name: job_id
        in: path
      - required: true
        schema:
          title: File Id
          type: integer
        example: 1
        name: file_id
        in: path
      - required: true
        schema:
          title: Revision
          type: string
        example: latest
        name: revision
        in: path
      - required: false
        schema:
          title: Page Numbers
          minimum: 1.0
          minItems: 1
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
        example:
        - 1
        - 3
        - 4
        name: page_numbers
        in: query
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticularRevisionSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
  /annotation/{job_id}/{file_id}/changes/{revision}:
    get:
      tags:
      - Annotation
      - Revision
      - Annotation
      summary: Get annotation for latest or particular revision.
      operationId: get_annotation_for_given_revision_annotation__job_id___file_id__changes__revision__get
      parameters:
      - required: true
        schema:
          title: Job Id
          type: integer
        example: 1
        name: job_id
        in: path
      - required: true
        schema:
          title: File Id
          type: integer
        example: 1
        name: file_id
        in: path
      - required: true
        schema:
          title: Revision
          type: string
        example: latest
        name: revision
        in: path
      - required: true
        schema:
          title: X-Current-Tenant
          type: string
        example: test
        name: X-Current-Tenant
        in: header
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParticularRevisionSchema'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConnectionErrorSchema'
        '404':
          description: Not Found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotFoundErrorSchema'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - TenantDependencyDocs: []
...
components:
  schemas:
    AnnotatedDocSchema:
      title: AnnotatedDocSchema
      required:
      - revision
      - user
      - pipeline
      - date
      - file_id
      - job_id
      - pages
      - validated
      - failed_validation_pages
      - tenant
      type: object
      properties:
        revision:
          title: Revision
          type: string
          example: 20fe52cce6a632c6eb09fdc5b3e1594f926eea69
        user:
          title: User
          type: string
          format: uuid
          example: 0b0ea570-e4e8-4664-84ac-dd1122471fc5
        pipeline:
          title: Pipeline
          type: integer
          example: 1
        date:
          title: Date
          type: string
          format: date-time
          example: '2021-10-19 01:01:01'
        file_id:
          title: File Id
          type: integer
          example: 1
        job_id:
          title: Job Id
          type: integer
          example: 1
        pages:
          title: Pages
          type: object
          additionalProperties:
            type: string
          example:
            '1': 19fe52cce6a632c6eb09fdc5b3e1594f926eea69
            '2': adda414648714f01c1c9657646b72ebb4433c8b5
        validated:
          title: Validated
          minimum: 1.0
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
          example:
          - 1
          - 2
          - 10
        failed_validation_pages:
          title: Failed Validation Pages
          minimum: 1.0
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
          example:
          - 3
          - 4
        tenant:
          title: Tenant
          type: string
          example: badger-doc
        task_id:
          title: Task Id
          type: integer
          example: 2
    BadRequestErrorSchema:
      title: BadRequestErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: Bad request.
...
    ConnectionErrorSchema:
      title: ConnectionErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: 'Error: Connection error.'
    DocForSaveSchema:
      title: DocForSaveSchema
      type: object
      properties:
        base_revision:
          title: Base Revision
          type: string
          example: 20fe52cce6a632c6eb09fdc5b3e1594f926eea69
        user:
          title: User
          type: string
          format: uuid
          example: b0ac6d8c-7b31-4570-a634-c92b07c9e566
        pipeline:
          title: Pipeline
          type: integer
          example: 1
        pages:
          title: Pages
          type: array
          items:
            $ref: '#/components/schemas/PageSchema'
        validated:
          title: Validated
          minimum: 1.0
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
          example:
          - 1
          - 2
          - 10
        failed_validation_pages:
          title: Failed Validation Pages
          minimum: 1.0
          uniqueItems: true
          type: array
          items:
            minimum: 1.0
            type: integer
          example:
          - 3
          - 4
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    JobOutSchema:
      title: JobOutSchema
      required:
      - job_id
      - is_manual
      type: object
      properties:
        job_id:
          title: Job Id
          type: integer
          example: 1
        is_manual:
          title: Is Manual
          type: boolean
          example: true
...
    NotFoundErrorSchema:
      title: NotFoundErrorSchema
      required:
      - detail
      type: object
      properties:
        detail:
          title: Detail
          type: string
      example:
        detail: Resource was not found.
...
    ValidationError:
      title: ValidationError
      required:
      - loc
      - msg
      - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            type: string
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
...
  securitySchemes:
    TenantDependencyDocs:
      type: http
      scheme: bearer
...